Extract the judgment commit contract into its own kernel module
nauro_core.operations.commit_plan was both the contract and a facade
that reached its two phase modules through sixteen function-local
imports, while each phase module imported the facade back as its
namespace. The contract types now live in _commit_contract.py together
with the derivations the self-validating models recompute (payload
parsing, provenance, the claim contract, snapshot bytes, claim keys,
observation validation and the plan record), because a model that
rejects divergence from a derivation cannot sit below it. The phase
modules keep only their orchestration and import the contract directly,
and commit_plan.py is a top-level re-export of the public surface.

Import paths and pickle identity are unchanged. The kernel tests pin
each helper on its owning module with the same signatures and
docstrings as before, and the model-shape digest is unchanged once
module names normalize to the facade. The hosted consumer snapshot
regenerates only the module qualifier of the moved classes.
